requirements.txt
asgiref==3.7.2
boto3==1.34.7
botocore==1.34.7
dj-database-url==2.1.0
Django==4.2.10
django-browser-reload==1.12.1
django-storages==1.14.2
django-tailwind==3.8.0
gunicorn==21.2.0
jmespath==1.0.1
openai==1.6.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==7.6.0
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.5.0
